+++ lib/get_it.dart
@@ -150,10 +150,12 @@ abstract class GetIt {
   /// If you really need to you can disable the asserts by setting[allowReassignment]= true
   bool allowReassignment = false;
 
-  /// retrieves or creates an instance of a registered type [T] or [type] depending on the registration
+  /// retrieves or creates an instance of a registered type [T] depending on the registration
   /// function used for this type or based on a name.
   /// for factories you can pass up to 2 parameters [param1,param2] they have to match the types
   /// given at registration with [registerFactoryParam()]
+  /// [type] if you want to get an instance by a Type object instead of a generic parameter.This should
+  /// rarely be needed but can be useful if you have a runtime type and want to get an instance
   T get<T extends Object>({
     String? instanceName,
     dynamic param1,
@@ -165,6 +167,8 @@ abstract class GetIt {
   /// not ready with its initialization.
   /// for async factories you can pass up to 2 parameters [param1,param2] they have to match the
   /// types given at registration with [registerFactoryParamAsync()]
+  /// [type] if you want to get an instance by a Type object instead of a generic parameter.This should
+  /// rarely be needed but can be useful if you have a runtime type and want to get an instance
   Future<T> getAsync<T extends Object>({
     String? instanceName,
     dynamic param1,